added pre-casp16 uniref90 database

+++ download_template_database_and_tools.py
@@ -21,15 +21,18 @@ def rm_if_exists(directory):
         else:
             os.system("rm "+directory)
 
-def direct_download(tool, address, tools_dir):  ####Tools don't need to be configured after downloading and configuring
+def direct_download(tool, address, tools_dir, unzip=True):  ####Tools don't need to be configured after downloading and configuring
     os.chdir(tools_dir)
     tool_dir = os.path.join(tools_dir, tool)
     if not os.path.exists(tool_dir):
         rm_if_exists(tool_dir)
         os.system("wget "+address)
         print("Decompressing "+tool_dir)
-        os.system("tar -zxf "+tool+".tar.gz && rm "+tool+".tar.gz")
-        os.system("chmod -R 755 "+tool_dir)
+        if unzip:
+            os.system("tar -zxf "+tool+".tar.gz && rm "+tool+".tar.gz")
+            os.system("chmod -R 755 "+tool_dir)
+        else:
+            os.system("chmod -R 755 "+tool_dir)
         print("Downloading "+tool_dir+"....Done")
     else:
         print(tool+" has been installed "+tool_dir+"....Skip....")
@@ -59,5 +62,12 @@ if __name__ == '__main__':
         if os.path.exists(os.path.join(database_dir, db)):
             continue
         direct_download(db,"http://sysbio.rnet.missouri.edu/multicom_cluster/multicom3_db_tools/databases/"+db+".tar.gz",database_dir)
+
+    db_lst = ["uniref90_Pre-CASP16.fasta"]
+    for db in db_lst:
+        print("Download "+db)
+        if os.path.exists(os.path.join(database_dir, db)):
+            continue
+        direct_download(db,"http://sysbio.rnet.missouri.edu/multicom_cluster/multicom3_db_tools/databases/"+db,database_dir, unzip=False)
 
     print("\nConfiguration....Done")
